// packages/test/generatedCode/models/bookingBusiness.ts
import {BookingAppointment, BookingCustomerBase, BookingCustomQuestion, BookingSchedulingPolicy, BookingService, BookingStaffMemberBase, BookingWorkHours, Entity, PhysicalAddress} from './index';
import {Parsable} from '@microsoft/kiota-abstractions';

export interface BookingBusiness extends Entity, Partial<Parsable> {
    /** The street address of the business. The address property, together with phone and webSiteUrl, appear in the footer of a business scheduling page. The attribute type of physicalAddress is not supported in v1.0. Internally we map the addresses to the type others. */
    address?: PhysicalAddress;
    /** All the appointments of this business. Read-only. Nullable. */
    appointments?: BookingAppointment[];
    /** The hours of operation for the business. */
    businessHours?: BookingWorkHours[];
    /** The type of business. */
    businessType?: string;
    /** The set of appointments of this business in a specified date range. Read-only. Nullable. */
    calendarView?: BookingAppointment[];
    /** All the customers of this business. Read-only. Nullable. */
    customers?: BookingCustomerBase[];
    /** All the custom questions of this business. Read-only. Nullable. */
    customQuestions?: BookingCustomQuestion[];
    /** The code for the currency that the business operates in on Microsoft Bookings. */
    defaultCurrencyIso?: string;
    /** The name of the business, which interfaces with customers. This name appears at the top of the business scheduling page. */
    displayName?: string;
    /** The email address for the business. */
    email?: string;
    /** The scheduling page has been made available to external customers. Use the publish and unpublish actions to set this property. Read-only. */
    isPublished?: boolean;
    /** The languageTag property */
    languageTag?: string;
    /** The telephone number for the business. The phone property, together with address and webSiteUrl, appear in the footer of a business scheduling page. */
    phone?: string;
    /** The URL for the scheduling page, which is set after you publish or unpublish the page. Read-only. */
    publicUrl?: string;
    /** Specifies how bookings can be created for this business. */
    schedulingPolicy?: BookingSchedulingPolicy;
    /** All the services offered by this business. Read-only. Nullable. */
    services?: BookingService[];
    /** All the staff members that provide services in this business. Read-only. Nullable. */
    staffMembers?: BookingStaffMemberBase[];
    /** The URL of the business web site. The webSiteUrl property, together with address, phone, appear in the footer of a business scheduling page. */
    webSiteUrl?: string;
}